Finalização do projeto

+++ src/main/java/br/com/fiap/checkpoint3/model/Profissional.java
@@ -1,5 +1,6 @@
 package br.com.fiap.checkpoint3.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -13,7 +14,7 @@ public class Profissional {
     private Long id;
     private String nome;
     private String especialidade;
-    private double valorHora;
+    private BigDecimal valorHora;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     
@@ -23,30 +24,35 @@ public class Profissional {
     public void setId(Long id) {
         this.id = id;
     }
+
     public String getNome() {
         return nome;
     }
     public void setNome(String nome) {
         this.nome = nome;
     }
+
     public String getEspecialidade() {
         return especialidade;
     }
     public void setEspecialidade(String especialidade) {
         this.especialidade = especialidade;
     }
-    public double getValorHora() {
+
+    public BigDecimal getValorHora() {
         return valorHora;
     }
-    public void setValorHora(double valorHora) {
+    public void setValorHora(BigDecimal valorHora) {
         this.valorHora = valorHora;
     }
+
     public LocalDateTime getCreatedAt() {
         return createdAt;
     }
     public void setCreatedAt(LocalDateTime createdAt) {
         this.createdAt = createdAt;
     }
+    
     public LocalDateTime getUpdatedAt() {
         return updatedAt;
     }
